destructure home feed components into platform specific parent components
