+++ apple/OmnivoreKit/Sources/App/Views/Home/HomeView.swift
@@ -1,0 +1,17 @@
+import SwiftUI
+
+struct HomeView: View {
+  @StateObject private var viewModel = HomeFeedViewModel()
+
+  var body: some View {
+    #if os(iOS)
+      if UIDevice.isIPhone {
+        CompactHomeView(viewModel: viewModel)
+      } else {
+        HomeFeedContainerView(viewModel: viewModel)
+      }
+    #elseif os(macOS)
+      HomeFeedView(viewModel: viewModel)
+    #endif
+  }
+}
